Setup Drivetrain

+++ src/main/cpp/util/Limelight.cpp
@@ -48,17 +48,17 @@ Limelight::LEDMode Limelight::GetLEDMode() const
 }
 
 
-double Limelight::GetHorizontalOffset()
+float Limelight::GetHorizontalOffset()
 {
     return table_->GetNumber("tx", 0.0);
 }
 
-double Limelight::GetVerticalOffset()
+float Limelight::GetVerticalOffset()
 {
     return table_->GetNumber("ty", 0.0);
 }
 
-double Limelight::GetTargetArea()
+float Limelight::GetTargetArea()
 {
     return table_->GetNumber("ta", 0.0);
 }
@@ -71,7 +71,7 @@ bool Limelight::HasActiveTarget()
 units::meter_t Limelight::EstimateTargetDistance(units::degree_t camera_mounting_angle, units::degree_t y_offset_to_target,
                        units::meter_t camera_height, units::meter_t target_height)
 {
-    return (target_height - camera_height) / std::tan((camera_mounting_angle + y_offset_to_target).to<double>());
+    return (target_height - camera_height) / std::tan((camera_mounting_angle + y_offset_to_target).to<float>());
 }
 
 }
