+++ src/main/include/util/Tunable.h
@@ -10,18 +10,18 @@ class Tunable
 {
 public:
     // Initializes SmartDashboard values.
-    Tunable(const char* name, const double& value);
+    Tunable(const char* name, const float& value);
 
     // Non-const getter; updates value with respect to SmartDashboard and then returns that value.
-    double Get();
+    float Get();
 
     // Manually change the tunable variable and update SmartDashboard.
-    void Set(const double& newState);
+    void Set(const float& newState);
 
     ~Tunable() {}
 private:
     const char* name_;
-    double value_;
+    float value_;
 };
 
 }
